Final_proyect/src/requirements_backend.txt
fastapi
sqlalchemy
pydantic
pydantic-settings
uvicorn
psycopg2-binary
python-dotenv
